+++ test/utils_test.dart
@@ -24,44 +24,44 @@ void main() {
   group('Utils', () {
     test('obfuscatePassword', () {
       expect(
-          SpongeUtils.obfuscatePassword(
+          SpongeClientUtils.obfuscatePassword(
               '{"username":"test","password":"secret"}'),
           equals('{"username":"test","password":"***"}'));
-      expect(SpongeUtils.obfuscatePassword('{"username":null,"password":null}'),
+      expect(SpongeClientUtils.obfuscatePassword('{"username":null,"password":null}'),
           equals('{"username":null,"password":null}'));
       expect(
-          SpongeUtils.obfuscatePassword(
+          SpongeClientUtils.obfuscatePassword(
               '{"username":null,"password":null,"object":{"name":"value"}}'),
           equals(
               '{"username":null,"password":null,"object":{"name":"value"}}'));
       expect(
-          SpongeUtils.obfuscatePassword(
+          SpongeClientUtils.obfuscatePassword(
               '{"id":null,"username":"test","password":"password","authToken":null,"name":"TestAction","args":["TEST",null],"version":null}'),
           equals(
               '{"id":null,"username":"test","password":"***","authToken":null,"name":"TestAction","args":["TEST",null],"version":null}'));
 
-      expect(SpongeUtils.obfuscatePassword('{"passwordRetype":"secret!!!"}'),
+      expect(SpongeClientUtils.obfuscatePassword('{"passwordRetype":"secret!!!"}'),
           equals('{"passwordRetype":"***"}'));
-      expect(SpongeUtils.obfuscatePassword('{"PASSwordRetype":"secret!!!"}'),
+      expect(SpongeClientUtils.obfuscatePassword('{"PASSwordRetype":"secret!!!"}'),
           equals('{"PASSwordRetype":"***"}'));
-      expect(SpongeUtils.obfuscatePassword('{"a":"b"}'), equals('{"a":"b"}'));
+      expect(SpongeClientUtils.obfuscatePassword('{"a":"b"}'), equals('{"a":"b"}'));
     });
     test('isHttpSuccess', () {
-      expect(SpongeUtils.isHttpSuccess(200), isTrue);
-      expect(SpongeUtils.isHttpSuccess(404), isFalse);
+      expect(SpongeClientUtils.isHttpSuccess(200), isTrue);
+      expect(SpongeClientUtils.isHttpSuccess(404), isFalse);
     });
     test('isServerVersionCompatible', () {
       expect(
-          SpongeUtils.isServerVersionCompatible(
+          SpongeClientUtils.isServerVersionCompatible(
               '${SpongeClientConstants.SUPPORTED_SPONGE_VERSION_MAJOR_MINOR}.5'),
           isTrue);
-      expect(SpongeUtils.isServerVersionCompatible('0.4.3'), isFalse);
+      expect(SpongeClientUtils.isServerVersionCompatible('0.4.3'), isFalse);
     });
     test('formatIsoDateTimeZone', () async {
       await initializeTimeZone();
       var dateTimeZone = TZDateTime.from(
           DateTime.parse('2019-02-07T15:16:17'), getLocation('Europe/Paris'));
-      expect(SpongeUtils.formatIsoDateTimeZone(dateTimeZone),
+      expect(SpongeClientUtils.formatIsoDateTimeZone(dateTimeZone),
           equals('2019-02-07T15:16:17.000+01:00[Europe/Paris]'));
     });
     test('parseIsoDateTimeZone', () async {
@@ -69,7 +69,7 @@ void main() {
       var dateTimeZone = TZDateTime.from(
           DateTime.parse('2019-02-07T15:16:17'), getLocation('Europe/Paris'));
       expect(
-          SpongeUtils.parseIsoDateTimeZone(
+          SpongeClientUtils.parseIsoDateTimeZone(
               '2019-02-07T15:16:17.000+01:00[Europe/Paris]'),
           equals(dateTimeZone));
     });
@@ -84,8 +84,8 @@ void main() {
 
       var exampleMap = {for (var name in fieldNames) name: null};
 
-      expect(SpongeUtils.isAnnotatedValueMap(exampleMap), isTrue);
-      expect(SpongeUtils.isAnnotatedValueMap({'value': null}), isFalse);
+      expect(SpongeClientUtils.isAnnotatedValueMap(exampleMap), isTrue);
+      expect(SpongeClientUtils.isAnnotatedValueMap({'value': null}), isFalse);
     });
   });
 }